2.5.0 keyspace oome (#154)
* Put Errors in ex-info cause rather than data

This prevents duplication of data, since an ExceptionInfo instance always
contains two copies of its data. Nesting errors in the data section leads to
exponential (order 2) growth of the size of the exception.

Inserting the error in cause makes nested errors accessable through the `:via`
tag of ExceptionInfo.

* Formatting

* Keep only the latest prws error from sequential monad bindings.

So instead of there being an error for each bind call, there is only one for the
bind that caused the error. This is then passed back up the binding chain.

* Added prws-error?