don't include printed stacktrace in error variants (#1102)
The `decode-error-value` fn prints the exception to a string and includes it in the variant as error message. Issues with this approach:

* error messages can leak to clients (potentially including secrets/keys etc)
* massive strings (e.g. when it's large schema error) are not helpful in practice
* it can be an expensive operation

This PR replaces the printed exception with an opaque error message (the exception message can still leak something to the client, so we want to be cautious). But the exception itself still is accessible in the variant's value metadata (under `:exception`). Special handling for schema errors to provide _some_ info in the message and not break existing tests.