sales games 01 - prep (#1116)
foundational changes to smooth the way for a sales-games update

### [promisespromises](https://github.com/yapsterapp/yapster/pull/1116/commits/75f4fe5dfde24310816c9313af850705ae786ccc) 
- make `prpr.stream/realize-stream` into a macro - otherwise the contained `stream/reduce` got zero context information and errors were impossible to trace. even with this change, the `stream/reduce` only gets filename context and no line#/col# (because 2 levels of macroexpansion) so additionally, deprecate `prpr.stream/realize-stream` - a `(->> ... (stream/reduce conj []))` is considerable more functional
- introduce `prpr.promise.retry/retry` with obvious purpose
- introduce `prpr.stream/error-take!` a version of `take!` which respects `StreamError`s
- add a multi-arity version of `stream/map`
- add `prpr.stream/concat-seq-s` to work around memory leaks with `stream/transform` (this is probably the same leak which borked the cassandra-driver update for a while - the solution to that issue was to not use `stream/transform`)

### [postgresql name-quoting](https://github.com/yapsterapp/yapster/pull/1116/commits/58ab5da5d4d2fa208971a044561a7981e4655739)
- in the social-games module we create tables to receive customer data which often have table-names and column-names requiring quoting - this helper has reduced query boilerplate considerably

### [http connector improvements](https://github.com/yapsterapp/yapster/pull/1116/commits/b5ee84ac198120aeef6676bc970a801140937159)
the social-games client uses the yapster API quite hard, and exposed some weaknesses in the existing client which are corrected here
- more response formats are dealt with - aleph can be inscrutable about what type of object is chosen for a response body
- a `nil` http-client impl is provided - for those (perhaps most) situations where the default client is fine
- error reporting is improved - a data description of the failed request is added to the thrown exception
- a `retry` operation is provided, using the data-description of a failed request to make the request again as specified

`retry` is particularly useful for the social-games client - API errors sometimes happen, and, now that all the social-games messages are idempotent, retry makes running the client a lot more reliable

### [slurp-url improvements](https://github.com/yapsterapp/yapster/pull/1116/commits/0be9807787104512aa84c1a7772fc261470ef752)
social-games makes heavy use of `slurp-url` for uniformly retrieving files from local fs, S3 and jar-resources
- fix a memory leak
- introduce a `nil` implementation for easier testing
- add a `slurp-to-tempfile` fn

### [csv parser improvements](https://github.com/yapsterapp/yapster/pull/1116/commits/7ff034bf40c004bb39ac55e7f3df7ac8fab78c05)
social-games makes heavy use of CSV for customer data, and often directly from an `InputStream` slurped from S3. some issues were discovered
- i/o exceptions are now surfaced properly
- text inside parentheses is no longer removed from column names (broke VE data)
- `clojure-csv` parser options can now be passed through

### [model.test additions](https://github.com/yapsterapp/yapster/pull/1116/commits/35b9056fb84ef168f91fe4acb8c1be193e93a83c)
adds support for testing with an `ISlurpUrl` slurper

### [a new file abstraction](https://github.com/yapsterapp/yapster/pull/1116/commits/e4b87ae256cfa3f9b59b4aa86c3019524771257c)
social-games uses libraries of resource files which may be specialized for each customer, and moves files from remote to local and back again for customer-data processing pipeline. there was no simple API for these operations, so such an API is introduced here, providing :
- `copy`
- `move`
- `delete`
- `exists?`
- `list`
- `upload-tempfile`
for accessing the content of files, `slurp-url` is used

### [spreadsheet parsing](https://github.com/yapsterapp/yapster/pull/1116/commits/ffd81858edaef13053ecb0425f9eb443b59efb40)
we do not have control of the format customers give us data, and it may come in CSVs or it may come in excel spreadsheets. this change creates a new `parse-spreadsheet` multimethod which uniformly parses CSVs and excel files from local or remote locations (with `slurp-url` used to open the files)

`parse-spreadsheet` has the same interface as the existing `parse-delimited-text`, and all existing uses of `parse-delimited-text` are replaced with `parse-spreadsheet`, so we should get the ability to accept excel files everywhere we currently accept CSVs